third_party/blink/web_tests/external/wpt/navigation-api/ordering-and-transition/navigate-intercept_currententrychange-expected.txt

This is a testharness.js-based test.
[FAIL] event and promise ordering for same-document navigation.navigate() intercepted by intercept()
  assert_array_equals: expected property 3 to be "navigatesuccess" but got "committed fulfilled" (expected array ["navigate", "currententrychange", "handler run", "navigatesuccess", "committed fulfilled", "transition.committed fulfilled", "promise microtask", "finished fulfilled", "transition.finished fulfilled"] got ["navigate", "currententrychange", "handler run", "committed fulfilled", "transition.committed fulfilled", "promise microtask", "navigatesuccess", "finished fulfilled", "transition.finished fulfilled"])
Harness: the test ran to completion.
